Project is fully functional
Images & videos are renamed according to the date of creation with a counter at the end and a prefix in the front.

If a folder is not chosen, let the user know at the bottom. Require the user to select a folder, file type and time format. If those are not selected, throw an error message. If they are selected, show a green tick box on the right side of the buttons. When the rename button is clicked, show a success message.